add SignoutManager + remove bcast channel again
