+++ package.json
@@ -55,7 +55,6 @@
     "@react-navigation/native": "6.0.2",
     "@react-navigation/stack": "6.0.7",
     "babel-plugin-transform-remove-console": "^6.9.4",
-    "broadcast-channel": "^4.2.0",
     "dotenv": "^8.2.0",
     "electron-context-menu": "^2.3.0",
     "electron-log": "^4.3.5",
