+++ src/ONYXKEYS.js
@@ -131,4 +131,7 @@ export default {
 
     // Stores Workspace ID that will be tied to reimbursement account during setup
     REIMBURSEMENT_ACCOUNT_WORKSPACE_ID: 'reimbursementAccountWorkspaceID',
+
+    // Notifies all tabs that they should sign out and clear storage.
+    SHOULD_SIGN_OUT: 'shouldSignOut',
 };
