--- src/components/SignoutManager/index.native.js
@@ -1,20 +1,0 @@
-let signoutCallback = () => {};
-
-/**
- * @param {Function} callback
- */
-function registerSignoutCallback(callback) {
-    signoutCallback = callback;
-}
-
-/**
- * @param {String} errorMessage
- */
-function signOut(errorMessage) {
-    signoutCallback(errorMessage);
-}
-
-export default {
-    signOut,
-    registerSignoutCallback,
-};
